requirements.txt
django>=1.11,<2.0
django-extensions
django-filter>=1.0.1,<1.1
django-redis>=4.7.0,<4.8
djangorestframework>=3.6,<3.7
gunicorn>=19.6.0,<19.7
ipython>=5.2.2,<5.3
psycopg2>=2.6.2,<2.7